+++ src/main/java/com/xiaoleilu/hutool/http/Method.java
@@ -6,5 +6,5 @@ package com.xiaoleilu.hutool.http;
  *
  */
 public enum Method {
-	GET, POST, HEAD, OPTIONS, PUT, DELETE, TRACE, CONNECT;
+	GET, POST, HEAD, OPTIONS, PUT, DELETE, TRACE, CONNECT, PATCH;
 }